Refactoring done, (NEW) test.py added, source.py broken to multiple files.
