+++ main.py
@@ -1,0 +1,9 @@
+from gui import GUI
+
+def main() -> None:
+    gui_instance = GUI()
+    gui_instance.mainloop()
+
+if __name__ == '__main__':
+    main()
+
